Refactor: sync inbox
